prodigy task 2/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
mysqlclient 